Skip generating reports for empty resolves
Currently, when no resolve was run, but the user requested a report, ivy resolve will fail looking for the report file.

This changes that behavior so that if a resolve result has no associated artifacts, an html report won't be generated and if all the resolve results were empty, it will log a message to that effect to the console.

Testing Done:
Wrote tests to replicate the behavior, replicated the failure and made them pass.

CI away on the PR: 3114.

Bugs closed: 3108, 3114

Reviewed at https://rbcommons.com/s/twitter/r/3625/

Closes #3108, #3114