[Added tbr] Retain dependency ordering in ivy.xml files when provide_excludes present.
When we changed excludes handling to not modify jar objects, we changed from an OrderedDict to a dict and lost the consistent ordering.

This ensures that the ordering is consistent.

Testing Done:
Ran ivy tests locally. CI away on PR.

Bugs closed: 3241

Reviewed at https://rbcommons.com/s/twitter/r/3736/