Customize tarfile module next() method
tarfile.TarFile object is iterable and has a next() method. next() will parse the header and save parsed info. During parsing, a lot of checks are done, to make sure the header is valid. And if there is something wrong with the header, exceptions will be thrown. next() catches a lot of them but not reraise what it catches in all cases.

We have a corrupted tgz file inside Twitter. During parsing, an error in one of the headers is caught, but next() hide it silently. Thus cache read succeeds with no error. When pants uses this corrupted tgz for analysis, it throws exception because it couldn't find a file which is supposed to be in the tarball.

From source code (https://hg.python.org/cpython/file/2.7/Lib/tarfile.py#l2335), we can see that InvalidHeaderError will ONLY be raised if it happens in the beginning of the tar file. Actually, a lot of exceptions are hidden by tarfile module. tarfile module simply thinks these exceptions mark the end of tarball. It probably is not a violation of tar standard, but it is not acceptable for pants. Pants relies on the integrity of tarball. Any defect in tarball should cause cache read to fail. A partially extracted tarball serves no use to Pants.

This changeset does the following:
1. Customize next() method in tarfile.TarFile for Python2.x. Now it will throws exception loudly whenever an invalidheader is seen if errorlevel is bigger than 0. If on Python3.x interpreter, original tarfile.Tarfile will be used.
2. Add test cases.

Testing Done:
ci green:
https://travis-ci.org/pantsbuild/pants/builds/155570257

Bugs closed: 3730

Reviewed at https://rbcommons.com/s/twitter/r/4123/