Options fingerprinting as provided by `TaskIdentityFingerprintStrategy` has proven to be extremely useful. But since passing a `fingerprint_strategy` to `Task.invalidated` acts as an _override_ of the default strategy (rather than being additive), it is currently easy to accidentally disable options fingerprinting (as happened in #4177 when a `Task` began passing a strategy).
When options fingerprinting was initially introduced, we were less certain about its efficacy and risks, which likely led to the decision to make it very easy to disable fingerprinting Task options. But in hindsight, options fingerprinting is the right choice in all cases we can think of.

Make options fingerprinting very difficult to disable by pushing it down to the `CacheKeyGenerator`, which is initialized once per `Task`. Deprecate `TaskIdentityFingerprintStrategy`, as it is now redundant.

Options that are marked `fingerprint=True` will always be included in a `Task`'s fingerprint.

Squashed commit of the following:

commit 7abf3d168deb2051918b63a5399cb0e29af5b1dd
Author: Stu Hood <stuhood@gmail.com>
Date:   Thu Feb 16 13:16:45 2017 -0800

    Make cache_key_generator creation lazy.

commit 1d0464d9288f4c831e6fbc52c9ea74e7c3d29006
Author: Stu Hood <stuhood@gmail.com>
Date:   Thu Feb 16 12:52:51 2017 -0800

    Fix ivy task tests.

commit 43c7eef792fff065a267abf8107e00fa21f846be
Author: Stu Hood <stuhood@gmail.com>
Date:   Wed Feb 15 18:16:14 2017 -0800

    Oops

commit 02ec73c1119ee0f7524d406b5269617428c05b56
Author: Stu Hood <stuhood@gmail.com>
Date:   Wed Feb 15 17:57:34 2017 -0800

    Remove internal usages of TaskIdentityFingerprintStrategy.

commit ce27d46e0d5eb0355f1bfd303609fa664096ad9d
Author: Stu Hood <stuhood@gmail.com>
Date:   Wed Feb 15 13:24:39 2017 -0800

    Make options fingerprinting very difficult to disable by pushing it down to the CacheKeyGenerator